pyproject.toml
[tool.poetry]
name = "zoom-python-webhook"
version = "0.1.0"
description = "Fast API app for Zoom Webhook"
authors = ["Samuel Guillemet <samuel.guillemet@telecom-sudparis.eu>"]
readme = "README.md"
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.0"}
pydantic-settings = "^2.0.1"
pytz = "^2023.3"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
mypy = "^1.4.1"
mypy-extensions = "^1.0.0"
pylint = "^2.17.4"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pylint-pydantic = "^0.2.0"
pytest-env = "^0.8.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/SamuelGuillemet/zoom-python-webhook/issues"

[tool.isort]
profile = "black"

[tool.mypy]
no_strict_optional = true
ignore_missing_imports = true
